+++ mem/sram/sram.hh
@@ -26,8 +26,8 @@ class SRAM : public AbstractSRAM {
   SRAM(ObjectData &);
   ~SRAM();
 
-  void read(uint64_t, uint64_t, Event, uint64_t) override;
-  void write(uint64_t, uint64_t, Event, uint64_t) override;
+  void read(uint64_t, uint32_t, Event, uint64_t) override;
+  void write(uint64_t, uint32_t, Event, uint64_t) override;
 
   void createCheckpoint(std::ostream &) const noexcept override;
   void restoreCheckpoint(std::istream &) noexcept override;
